% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pafCoordsDotPlotly.R, R/seqplotter.R
\name{dotplotly_dotplot}
\alias{dotplotly_dotplot}
\alias{pafdotplot_make}
\alias{dotPlotr}
\alias{make_dotplot}
\title{A core function for plotting a paf.}
\usage{
dotplotly_dotplot(opt)

pafdotplot_make(
  inpaf_link,
  outplot_link,
  min_align = 11,
  min_query_aln = 11,
  keep_ref = 10000,
  similarity = T,
  h_lines = T,
  interactive = F,
  plot_size = 10,
  on_target = T,
  v = F,
  hllink = F,
  hltype = F
)

dotPlotr(seq1, seq2, wsize = 5, wstep = 1, nmatch = -1)

make_dotplot(targetfa, queryfa, wsize, outfile, save = T, debugmode = F)
}
\arguments{
\item{opt}{a list of parameters.}

\item{inpaf_link}{\link{character/link} a link to the paf to be plotted}

\item{outplot_link}{\link{character/link} a link to where the output pdf will be saved}

\item{min_align}{not sure. set to 11, keep it there.}

\item{min_query_aln}{not sure. set to 11, keep it there.}

\item{keep_ref}{number of sequences to keep or something. Dont touch.}

\item{similarity}{\link{T/F} indicate sequence similarity by color code (default: T)}

\item{h_lines}{\link{T/F} plot horizontal lines (default: T)}

\item{interactive}{\link{T/F} unsure. Dotplotly? Untested, keep at F.}

\item{plot_size}{\link{numeric} plot size in inch.}

\item{on_target}{\link{T/F} dont remember. default: T}

\item{v}{\link{T/F} unsure. default: F}

\item{hllink}{\link{character/link} link to an SD annotation file (bed, tsv or paf) to include as
highlights in the plot.}

\item{hltype}{\link{character} filetype of hllink. Can be 'NULL', 'bed', 'tsv', 'paf'.}

\item{seq1}{A DNA sequence, as character or DNAString}

\item{seq2}{A DNA sequence, as character or DNAString}

\item{save}{\link{bool} if T, save plot to a file. Otherwise, return.}

\item{seq1link}{\link{character/link} A link to a target fasta file, as character or DNAString}

\item{seq2link}{\link{character/link} A link to a query fasta file, as character or DNAString}
}
\value{
Nothing, but writes a pdf with a dotplot.

Nothing, but writes a pdf with a dotplot.

a ggplot2 object: exact dotplot between the two seqs.

a ggplot2 object or nothing, depending on the 'save' parameter.
}
\description{
This is extracted and an edited version of
the script "pafCoordsDotPlotly.R" from the dotplotly package
(https://github.com/tpoorten/dotPlotly) by Tom Poorten.
Should be run through the 'pafdotplot_make' wrapper function
which helps in selecting input parameters.

T

Give me two sequences (character of DNAString I think),
and I will return to you a ggplot object with an exact (bp-precise)
dotplot. This relies on the mkDotPlotDataFrame function from 'dotplot' package.

Give me two sequences (character of DNAString I think),
and I will return to you a ggplot object with an exact (bp-precise)
dotplot. This relies on the mkDotPlotDataFrame function from 'dotplot' package.
}
\author{
Tom Poorten, edited by Wolfram Höps

Wolfram Höps

Wolfram Höps

Wolfram Höps
}
