% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqmodder.R
\name{mutate_seq}
\alias{mutate_seq}
\alias{carry_out_inv}
\alias{sds_document_inversion}
\alias{carry_out_del}
\alias{sds_document_deletion}
\alias{expand_one_of_pair}
\alias{expand_whole_pair}
\alias{carry_out_dup}
\alias{sds_document_duplication}
\title{Introduce mutations to a DNA sequence.}
\usage{
mutate_seq(seq_fasta, sds_tsv, sv_instr_txt, outfasta, outsd, debug = F)

carry_out_inv(seq, sds, sv)

sds_document_inversion(sds, sv)

carry_out_del(seq, sds, sv)

sds_document_deletion(sds, sv)

expand_one_of_pair(sds, uid, n_of_pair, dupsize)

expand_whole_pair(sds, uid, dupsize)

carry_out_dup(seq, sds, sv)

sds_document_duplication(sds, sv)
}
\arguments{
\item{seq_fasta}{\link{character/link} A single-sequence fasta file containing the
original sequence, which is to be modified. Should contain SDs.}

\item{sds_tsv}{\link{character/link} A tsv file DESCRIBING the seq_fasta. This information will
be used to update presumed SD position post modification (not working well yet).}

\item{sv_instr_txt}{\link{character/link} A tab-separated file with colums \link{SD-uid, SV}.
SD-uid has to be existing in sds_tsv. SV can be \link{INV, DEL, DUP}. s}

\item{outfasta}{\link{character/link} Outputfile for the modified fasta}

\item{outsd}{\link{character/link} Outputfile for the modified SD tsv file.}

\item{seq}{\link{character} DNA sequence as character vector}

\item{sds}{\link{data.frame} A dataframe with tsv SD information}

\item{sv}{\link{data.frame} A data framewith columns \link{SD-uid, SV}.
SD-uid has to be existing in sds_tsv. SV can be \link{INV, DEL, DUP}}
}
\value{
nothing. But output files written.

a list object containing the mutated sequence and sds dataframe.

a mutated sds dataframe.

a list object containing the mutated sequence and sds dataframe.

a mutated sds dataframe.

a list object containing the mutated sequence and sds dataframe.

a mutated sds dataframe.
}
\description{
This is sitting at the core of the nahrchainer module.

This function is typically called by the mutate_seq function

This function is typically called by the carry_out_inv function.
ALSO the function is highly experimental and probably not working.
We recommend not rely on it at this point.

This function is typically called by the mutate_seq function

This function is typically called by the carry_out_del function.
ALSO the function is highly experimental and probably not working.
We recommend not rely on it at this point.

This function is typically called by the mutate_seq function

This function is typically called by the carry_out_dup function.
ALSO the function is highly experimental and probably not working.
We recommend not rely on it at this point.
}
\author{
Wolfram Höps

Wolfram Höps

Wolfram Höps

Wolfram Höps

Wolfram Höps

Wolfram Höps

Wolfram Höps

Wolfram Höps

Wolfram Höps
}
